Test cases created
