+++ src/test/java/CatTest.java
@@ -1,9 +1,7 @@
 import com.example.Cat;
-import com.example.Feline;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
 
@@ -12,10 +10,7 @@ import java.util.List;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(MockitoJUnitRunner.class)
-public class CatTest {
-    @Mock
-    static Feline feline;
-
+public class CatTest extends BaseTest {
     @Test
     public void catHasSound() {
         Cat cat = new Cat(feline);
