+++ src/test/java/LionTest.java
@@ -1,6 +1,4 @@
-import com.example.Feline;
 import com.example.Lion;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
@@ -11,17 +9,9 @@ import java.util.List;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(Parameterized.class)
-public class LionTest {
-
-    @Before
-    public void init() {
-        MockitoAnnotations.openMocks(this);
-    }
+public class LionTest extends BaseTest {
     private final String sex;
     private final Boolean mane;
-
-    @Mock
-    Feline feline;
 
     public LionTest(String sex, Boolean mane) {
         this.sex = sex;
@@ -32,26 +22,39 @@ public class LionTest {
     public static Object[][] getTestData() {
         return new Object[][] {
                 {"Самец", true},
-                {"Самка", false}
+                {"Самка", false},
+                {"test", false}
         };
     }
 
     @Test
-    public void testSex() throws Exception {
-        Lion lion = new Lion(sex, feline);
-        assertEquals(mane, lion.doesHaveMane());
+    public void checkIfLionHasMane() {
+        try {
+            Lion lion = new Lion(sex, feline);
+            assertEquals(mane, lion.doesHaveMane());
+        } catch (Exception e) {
+            assertEquals("Используйте допустимые значения пола животного - самец или самка", e.getMessage());
+        }
     }
     @Test
-    public void haveFiveKittens() throws Exception {
-        Lion lion = new Lion(sex, feline);
+    public void checkCountOfKittens() {
+        try {
+            Lion lion = new Lion(sex, feline);
 
-        Mockito.when(feline.getKittens()).thenReturn(5);
-        assertEquals(5, lion.getKittens());
+            Mockito.when(feline.getKittens()).thenReturn(5);
+            assertEquals(5, lion.getKittens());
+        } catch (Exception e) {
+            assertEquals("Используйте допустимые значения пола животного - самец или самка", e.getMessage());
+        }
     }
     @Test
-    public void getFood() throws Exception {
-        Lion lion = new Lion(sex, feline);
-        Mockito.when(feline.getFood("Хищник")).thenReturn(List.of("Мясо"));
-        assertEquals(List.of("Мясо"), lion.getFood());
+    public void getFood() {
+        try {
+            Lion lion = new Lion(sex, feline);
+            Mockito.when(feline.getFood("Хищник")).thenReturn(List.of("Мясо"));
+            assertEquals(List.of("Мясо"), lion.getFood());
+        } catch (Exception e) {
+            assertEquals("Используйте допустимые значения пола животного - самец или самка", e.getMessage());
+        }
     }
 }
